Allow the caller of docker-compose to provide binaries via volumes
This is a small step towards making the running of PWD more efficient.
Building the parts of PWD outside of a container is always going to be
far more efficient than inside (and saves mapping of caches etc).

As a later step we could move entirely to this approach and use a
smaller image for running PWD.